Converted everything to ES6 for REST API, updated dependency for Vue CLI. Also added Swagger dependency, will add doc later
