+++ api/swagger.js
@@ -1,0 +1,1 @@
+import swaggerUi from 'swagger-ui-express';
